Use FIPS compliant settings and validation
