+++ internal/pkg/agent/vault/const_fips.go
@@ -1,0 +1,12 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
+
+//go:build requirefips
+
+package vault
+
+const (
+	// saltSize in FIPS must be at least 128 bits
+	saltSize = 16
+)
