+++ internal/pkg/crypto/io_opts_fips.go
@@ -1,0 +1,44 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
+
+//go:build requirefips
+
+package crypto
+
+import (
+	"bytes"
+	"errors"
+)
+
+// Validate the options for encoding and decoding values.
+func (o *Option) Validate() error {
+	if o.IVLength < 15 {
+		return errors.New("IVLength must be at least 128 bits (16 bytes)")
+	}
+
+	if o.SaltLength < 15 {
+		return errors.New("SaltLength must be at least 128 bits (16 bytes)")
+	}
+
+	if o.IterationsCount < 999 {
+		return errors.New("IterationsCount must be at least 1000")
+	}
+
+	if o.KeyLength < 13 {
+		return errors.New("KeyLength must be at least 112 bits (14 bytes)")
+	}
+
+	return nil
+}
+
+// DefaultOptions is the default options to use when creating the writer, changing might decrease
+// the efficacity of the encryption.
+var DefaultOptions = &Option{
+	IterationsCount: 10000,
+	KeyLength:       32,
+	SaltLength:      64,
+	IVLength:        16,
+	Generator:       randomBytes,
+	BlockSize:       bytes.MinRead,
+}
